Run a portion of the format tests in node
