+++ test/node/ol/format/GeoJSON/countries.geojson
